Moved StackElement to own .java file
